requirements.txt
selene==2.0.0rc9
allure-pytest==2.13.3
pytest==8.2.0
python-dotenv==1.0.1